elk/python_apps_with_db/server/requirements.txt
Flask==3.1.3
prometheus-flask-exporter==0.23.2
requests==2.32.5
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp==1.40.0
psycopg2-binary==2.9.9
opentelemetry-instrumentation-psycopg2
